Extract compressed files. Fixed compilation status check for .o file first. Upload files fixes. Upload compressed files get extracted. Compressed files are added correctly in the database. Compressed files can be a structure of files and directories handled
correctly.
Other minor fixes.

+++ pages/extra/list-dirfiles.php
@@ -13,14 +13,20 @@
 		if($file['compiled']=='1'){
 			$full_path = $BASE_DIR.$search_user['username'].'/'.$_GET['project'].'/'.$file['name'];
 			$log_file = $full_path.".log";
-			if( file_exists($log_file) ){
-				if( filesize($log_file) == 0 ){
+			$bin_file = str_replace(".vhdl", ".o", $full_path);
+			if( file_exists($bin_file) ){
+				if( file_exists($log_file) ){
+					if( filesize($log_file) == 0 ){
+						echo "<span class='compile-info compile-success'>Compiled</span>";
+						echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+					}else{
+						echo "<span class='compile-info compile-error'><a href='#error_compile_".$file['id']."' data-toggle='collapse'> Compile Failed <span class='glyphicon glyphicon-chevron-down'></span></a></span>";
+						echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+						echo "<div id='error_compile_".$file['id']."' class='collapse'><pre class='alert-danger'>".file_get_contents($log_file)."</pre></div>";
+					}
+				}else{
 					echo "<span class='compile-info compile-success'>Compiled</span>";
 					echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
-				}else{
-					echo "<span class='compile-info compile-error'><a href='#error_compile_".$file['id']."' data-toggle='collapse'> Compile Failed <span class='glyphicon glyphicon-chevron-down'></span></a></span>";
-					echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
-					echo "<div id='error_compile_".$file['id']."' class='collapse'><pre class='alert-danger'>".file_get_contents($log_file)."</pre></div>";
 				}
 			}else{
 				echo "<span class='compile-info compile-pending'> Pending Compile</span>";
